# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AgentFlow API
  version: 0.1.0
  description: |
    HTTP surface for AgentFlow runs, agents, and streaming events.
    Behavioral notes and SSE event types are documented in docs/api-contract.md.
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8000
    description: Local development (Java API default port)
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/runs:
    get:
      tags: [Runs]
      summary: List runs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Runs ordered by created_at descending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
    post:
      tags: [Runs]
      summary: Create run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCreateRequest"
      responses:
        "202":
          description: Run accepted (pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/runs/{id}:
    get:
      tags: [Runs]
      summary: Get run
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run with steps, message preview, checkpoints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/runs/{id}/messages:
    get:
      tags: [Runs]
      summary: Page run messages
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: cursor
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Message page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePage"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/runs/{id}/cancel:
    post:
      tags: [Runs]
      summary: Cancel run
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "204":
          description: Cancel signal sent
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/runs/{id}/retry:
    post:
      tags: [Runs]
      summary: Retry failed run
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRetryRequest"
      responses:
        "202":
          description: Run re-queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/runs/{id}/resume:
    post:
      tags: [Runs]
      summary: Resume run waiting for human approval
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunResumeRequest"
      responses:
        "202":
          description: Run re-queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API key
  parameters:
...
    RunId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    TooManyRequests:
      description: Agent quota exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
...
  schemas:
    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    RunStatus:
      type: string
      enum:
        - pending
        - running
        - succeeded
        - failed
        - cancelled
        - waiting_human
    RunCreateRequest:
      type: object
      required: [agent_id]
      properties:
        agent_id:
          type: string
        input:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        adapter:
          type: string
          nullable: true
    RunRetryRequest:
      type: object
      properties:
        checkpoint_index:
          type: integer
    RunResumeRequest:
      type: object
      properties:
        input:
          type: object
          additionalProperties: true
    RunUsage:
      type: object
      required:
        - tokens_in
        - tokens_out
        - cost_usd
        - step_count
        - failed_step_count
        - tool_call_count
        - failed_tool_call_count
      properties:
        tokens_in:
          type: integer
        tokens_out:
          type: integer
        cost_usd:
          type: number
          format: double
        latency_ms:
          type: integer
          nullable: true
        step_count:
          type: integer
        failed_step_count:
          type: integer
        tool_call_count:
          type: integer
        failed_tool_call_count:
          type: integer
    ToolCall:
      type: object
      required: [id, name, arguments]
      properties:
        id:
          type: string
        name:
          type: string
        arguments:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: string
          nullable: true
        latency_ms:
          type: integer
          nullable: true
    Step:
      type: object
      required:
        - id
        - index
        - node
        - status
        - input
        - tool_calls
        - created_at
        - updated_at
      properties:
        id:
          type: string
        index:
          type: integer
        node:
          type: string
        status:
          $ref: "#/components/schemas/RunStatus"
        input:
          type: object
          additionalProperties: true
        output:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: string
          nullable: true
        latency_ms:
          type: integer
          nullable: true
        tokens_in:
          type: integer
          nullable: true
        tokens_out:
          type: integer
          nullable: true
        cost_usd:
          type: number
          format: double
          nullable: true
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, index, role, content, extra, created_at]
      properties:
        id:
          type: string
        index:
          type: integer
        step_id:
          type: string
          nullable: true
        role:
          type: string
          enum: [system, user, assistant, tool]
        name:
          type: string
          nullable: true
        content:
          type: string
        tool_call_id:
          type: string
          nullable: true
        extra:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    MessagePage:
      type: object
      required: [items, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        next_cursor:
          type: integer
          nullable: true
        has_more:
          type: boolean
    Checkpoint:
      type: object
      required: [id, index, created_at]
      properties:
        id:
          type: string
        index:
          type: integer
        label:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    Run:
      type: object
      required:
        - id
        - tenant_id
        - agent_id
        - adapter
        - status
        - input
        - created_at
        - updated_at
        - steps
        - messages
        - checkpoints
        - usage
      properties:
        id:
          type: string
        tenant_id:
          type: string
        project_id:
          type: string
          nullable: true
        agent_id:
          type: string
        adapter:
          type: string
        status:
          $ref: "#/components/schemas/RunStatus"
        input:
          type: object
          additionalProperties: true
        output:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/Step"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        messages_truncated:
          type: boolean
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/Checkpoint"
        usage:
          $ref: "#/components/schemas/RunUsage"
...